fix(hud): fix hud render bugs
- Fix blurry hud display when minmap enabled.
- Remove unused textures.
- Change textures for simpler display feeling.
- Correct unexpectedly clamped hp bar alpha value.

+++ src/gameplay/hud/ScoreDisplay.cc
@@ -14,7 +14,7 @@ void ScoreDisplay::draw() {
     if (t - beginTime > SCORE_VARY_TIME) {
         displayScore = targetScore;
     } else {
-        auto p = (unsigned int) ((t - beginTime) / double(SCORE_VARY_TIME));
+        auto p = ((t - beginTime) / double(SCORE_VARY_TIME));
         displayScore = p * targetScore + (1 - p) * beginScore;
         alpha = 1 - 0.3 * p;
     }
